pyproject.toml
[project]
name = "codex-ea"
version = "0.1.0"
description = "Use GPT-5 Codex programatically, using your existing subscription."
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["openai==1.108.0", "python-dotenv==1.1.1"]
